requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# AI/ML
google-generativeai==0.8.1
httpx==0.27.2

# Image Processing
pillow==10.4.0

# Environment Variables
python-dotenv==1.0.1

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Firebase Admin SDK (Backend Only)
firebase-admin==6.5.0

# Background Scheduling (for daily mandi price fetch)
apscheduler==3.10.4
